Add endpoint for dynamic javascript files
