+++ InvenTree/templates/base.html
@@ -114,6 +114,8 @@ InvenTree
 <script type='text/javascript' src="{% static 'script/inventree/notification.js' %}"></script>
 <script type='text/javascript' src="{% static 'script/inventree/sidenav.js' %}"></script>
 
+<script type='text/javascript' src="{% url 'part.js' %}"></script>
+
 <script type='text/javascript' src="{% static 'fontawesome/js/solid.js' %}"></script>
 <script type='text/javascript' src="{% static 'fontawesome/js/brands.js' %}"></script>
 <script type='text/javascript' src="{% static 'fontawesome/js/fontawesome.js' %}"></script>
